MNIST Pytorch/requirements.txt
fastapi>=0.68.0
uvicorn[standard]
torch>=1.9.0
torchvision>=0.10.0
Pillow